requirements.txt
pymupdf
pillow
openai
tqdm